move nested components to main folder
