--- client/src/components/Search/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Search";
