--- client/src/components/Sign In/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SignIn";
